+++ src/restfw/testing/__init__.py
@@ -3,6 +3,8 @@
 :Authors: cykooz
 :Date: 06.12.2016
 """
+import warnings
+
 try:
     import pytest
     pytest.register_assert_rewrite(
@@ -12,38 +14,27 @@ try:
 except (ImportError, AttributeError):
     pass
 
-from contextlib import contextmanager
-from pyramid.interfaces import IRequestFactory, IRootFactory
-from pyramid.request import Request, apply_request_extensions
-from pyramid.threadlocal import get_current_request, RequestContext
-from pyramid.traversal import DefaultRootFactory
-
 from .resource_testing import assert_resource
+from .. import utils
 
 
 def get_pyramid_root(request=None):
-    request = request or get_current_request()
-    if getattr(request, 'root', None) is None:
-        root_factory = request.registry.queryUtility(IRootFactory, default=DefaultRootFactory)
-        root = root_factory(request)  # Initialise pyramid root
-        request.root = root
-    return request.root
+    warnings.warn(
+        'Function restfw.testing:get_pyramid_root will be removed at next major release. '
+        'Please use it new version restfw.utils:get_pyramid_root',
+        stacklevel=2
+    )
+    return utils.get_pyramid_root(request)
 
 
-@contextmanager
-def open_pyramid_request(registry):
+def open_pyramid_request(app_config):
     """
-    :type registry: pyramid.registry.Registry
+    :type app_config: pyramid.config.Configurator
     :rtype: pyramid.request.Request
     """
-    request_factory = registry.queryUtility(IRequestFactory, default=Request)
-    request = request_factory.blank('http://localhost')
-    request.registry = registry
-    apply_request_extensions(request)
-    get_pyramid_root(request)
-    context = RequestContext(request)
-    context.begin()
-    try:
-        yield request
-    finally:
-        context.end()
+    warnings.warn(
+        'Function restfw.testing:open_pyramid_request will be removed at the next major release. '
+        'Please use it new version restfw.utils:open_pyramid_request',
+        stacklevel=2
+    )
+    return utils.open_pyramid_request(app_config.registry)